requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
cryptography==43.0.3
firebase-admin==6.5.0
gTTS==2.5.4
